// src/types/user.ts
export type UserRegister = {
  fullName: string;
  email: string;
  password: string;
};
export type UserLogin = {
  email: string;
  password: string;
};
export type UserUpdate = {
  fullName?: string;
  email?: string;
  password?: string;
  otp?: string | null;
  isVerified?: boolean;
  resetToken?: string | null;
  resetTokenExpires?: Date | null;
  role?: "USER" | "ADMIN";
};
export type User = {};